requirements.txt
streamlit
pandas
st-gsheets-connection
gspread
google-generativeai
openai
Pillow
python-dotenv